Progress save– separate `DistributionComparator` and `GraphComparator` classes.
Wanted to save my progress here. As I began writing out the specific implementation for `DistributionComparator` it became clear that this functionality for aggregating a list of topological descriptors should be handled in the `TopologicalDistance` objects. Going forward we'll have a single `Comparator` and the relevant distance classes will support an `aggregate` method to summarize the distribution with a single representative (in the paper we use average).